# voice-journal-api-spec.yaml
openapi: 3.0.3
info:
  title: "Voice Journal API"
  version: "1.0.0"
  description: |
    Voice Journal은 사용자가 매일 밤 내일의 계획을 음성으로 녹음하고,
    아침에 자신의 목소리로 들으며 동기부여를 받는 개인 성장 플랫폼입니다.
    
    ## 주요 기능
    - 🎤 음성 파일 업로드 및 OpenAI Whisper를 통한 텍스트 변환
    - 📝 계획(plan)과 회고(reflection) 기록 관리
    - 🎯 목표 설정 및 달성률 추적
    - 📊 성장 패턴 분석 및 인사이트 제공
    
    ## 인증 방식
    이 API는 JWT 기반 Bearer 토큰 인증을 사용합니다.
    모든 보호된 엔드포인트는 Authorization 헤더에 'Bearer {token}' 형식으로 토큰을 포함해야 합니다.
    
    ## 응답 형식
    모든 응답은 JSON 형식이며, 에러 발생 시 일관된 ErrorResponse 스키마를 따릅니다.
    
    ## 파일 업로드
    음성 파일 업로드는 S3 Pre-signed URL을 통해 이루어지며, 
    지원되는 형식은 MP3, WAV, M4A, OGG (최대 25MB)입니다.
  termsOfService: "https://voicejournal.com/terms"
  contact:
    name: "Voice Journal 지원팀"
    url: "https://voicejournal.com/support"
    email: "support@voicejournal.com"
  license:
    name: "MIT"
    url: "https://opensource.org/licenses/MIT"
servers:
  - url: "https://api.voicejournal.com"
    description: "Production server"
  - url: "https://staging-api.voicejournal.com"
    description: "Staging server for testing"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT 토큰을 Authorization 헤더에 'Bearer {token}' 형식으로 포함해주세요."
  schemas:
    ErrorResponse:
      type: object
      description: "표준 에러 응답 형식"
      properties:
        code:
          type: integer
          description: "HTTP 상태 코드"
          example: 400
        message:
          type: string
          description: "사용자에게 표시할 에러 메시지"
          example: "잘못된 요청입니다."
        details:
          type: string
          description: "개발자를 위한 상세 에러 정보 (선택적)"
          example: "filename 파라미터가 누락되었습니다."
        timestamp:
          type: string
          format: date-time
          description: "에러 발생 시간 (ISO 8601)"
          example: "2024-01-15T10:30:00Z"
        requestId:
          type: string
          description: "요청 추적을 위한 고유 ID (선택적)"
          example: "req_abc123def456"
      required: [code, message, timestamp]
...
    TranscriptionRequest:
      type: object
      description: "음성 변환 요청"
      properties:
        audioUrl:
          type: string
          format: uri
          description: "변환할 음성 파일의 URL"
          example: "https://cdn.voicejournal.com/temp/audio123.mp3"
        language:
          type: string
          description: "언어 코드 (선택적, 기본값: ko)"
          example: "ko"
        prompt:
          type: string
          description: "변환 정확도 향상을 위한 힌트 텍스트 (선택적)"
          example: "이것은 내일 계획에 대한 음성 메모입니다."
      required: [audioUrl]
    TranscriptionResponse:
      type: object
      description: "음성 변환 결과"
      properties:
        success:
          type: boolean
          description: "변환 성공 여부"
          example: true
        text:
          type: string
          description: "변환된 텍스트"
          example: "내일은 아침 7시에 일어나서 운동을 하겠습니다."
        language:
          type: string
          description: "감지된 언어 코드"
          example: "ko"
        duration:
          type: number
          format: float
          description: "음성 길이 (초)"
          example: 165.5
        confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: "전체 변환 신뢰도"
          example: 0.95
        segments:
          type: array
          description: "세그먼트별 상세 정보 (선택적)"
          items:
            type: object
            properties:
              start:
                type: number
                format: float
                description: "시작 시간 (초)"
                example: 0.0
              end:
                type: number
                format: float
                description: "종료 시간 (초)"
                example: 5.2
              text:
                type: string
                description: "세그먼트 텍스트"
                example: "내일은 아침 7시에"
              confidence:
                type: number
                format: float
                description: "세그먼트 신뢰도"
                example: 0.98
        processingTime:
          type: number
          format: float
          description: "처리 시간 (초)"
          example: 3.2
      required: [success, text]
...
    PresignedUrlResponse:
      type: object
      description: "S3 Pre-signed URL 응답"
      properties:
        uploadUrl:
          type: string
          format: uri
          description: "파일 업로드용 pre-signed URL"
          example: "https://s3.amazonaws.com/voice-journal-bucket/audio/user123/20240115_213000.mp3?X-Amz-Algorithm=AWS4-HMAC-SHA256&X-Amz-Credential=..."
        fileUrl:
          type: string
          format: uri
          description: "업로드 완료 후 파일 접근 URL"
          example: "https://cdn.voicejournal.com/audio/user123/20240115_213000.mp3"
        fileKey:
          type: string
          description: "S3 객체 키"
          example: "audio/user123/20240115_213000.mp3"
        expiresIn:
          type: integer
          description: "URL 만료까지 남은 시간 (초)"
          example: 3600
        maxFileSize:
          type: integer
          description: "최대 업로드 가능 파일 크기 (바이트)"
          example: 26214400
      required: [uploadUrl, fileUrl, fileKey, expiresIn]
paths:
...
  /api/v1/upload/presigned-url:
    get:
      tags: ["File Upload"]
      summary: "S3 Pre-signed URL 발급"
      description: |
        음성 파일 업로드를 위한 S3 pre-signed URL을 발급합니다.
        
        - 지원 형식: MP3, WAV, M4A, OGG, WEBM
        - 최대 파일 크기: 25MB (OpenAI Whisper 제한)
        - URL 유효 시간: 1시간
      operationId: "getPresignedUrl"
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: filename
          required: true
          schema:
            type: string
            pattern: "^[a-zA-Z0-9._-]+\\.(mp3|wav|m4a|ogg|webm)$"
            example: "morning_plan.mp3"
          description: "업로드할 파일명 (확장자 포함)"
        - in: query
          name: contentType
          required: true
          schema:
            type: string
            enum: ["audio/mpeg", "audio/wav", "audio/m4a", "audio/ogg", "audio/webm"]
            example: "audio/mpeg"
          description: "파일 MIME 타입"
        - in: query
          name: fileSize
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 26214400
            example: 2097152
          description: "파일 크기 (바이트, 선택적)"
      responses:
        '200':
          description: "Pre-signed URL 발급 성공"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresignedUrlResponse'
        '400':
          description: "잘못된 요청 (지원하지 않는 파일 형식, 크기 초과 등)"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: "인증 실패"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: "요청 한도 초과"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: "서버 내부 오류"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    options:
      tags: ["File Upload"]
      summary: "CORS preflight for presigned URL"
      description: "CORS preflight 요청을 처리합니다."
      operationId: "presignedUrlOptions"
      responses:
        '204':
          description: "CORS preflight 성공"
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: "*"
            Access-Control-Allow-Methods:
              schema:
                type: string
                example: "GET, OPTIONS"
            Access-Control-Allow-Headers:
              schema:
                type: string
                example: "Content-Type, Authorization"
            Access-Control-Max-Age:
              schema:
                type: integer
                example: 86400

  # Voice Processing
  /api/v1/transcribe:
    post:
      tags: ["Voice Processing"]
      summary: "음성 텍스트 변환"
      description: |
        OpenAI Whisper를 사용하여 음성 파일을 텍스트로 변환합니다.
        
        - 지원 언어: 한국어, 영어 등 99개 언어
        - 자동 언어 감지 또는 수동 지정 가능
        - 세그먼트별 타임스탬프 제공
        - 변환 신뢰도 점수 포함
      operationId: "transcribeAudio"
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TranscriptionRequest'
      responses:
        '200':
          description: "변환 성공"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranscriptionResponse'
        '400':
          description: "잘못된 요청 (파일 형식, 크기, URL 등)"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid_file:
                  summary: "지원하지 않는 파일"
                  value:
                    code: 400
                    message: "지원하지 않는 파일 형식입니다."
                    details: "MP3, WAV, M4A, OGG 파일만 지원됩니다."
                    timestamp: "2024-01-15T10:30:00Z"
                file_too_large:
                  summary: "파일 크기 초과"
                  value:
                    code: 400
                    message: "파일 크기가 너무 큽니다."
                    details: "최대 25MB까지 지원됩니다."
                    timestamp: "2024-01-15T10:30:00Z"
        '401':
          description: "인증 실패"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '402':
          description: "할당량 부족 (OpenAI API 크레딧 부족)"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: "요청 한도 초과"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: "서버 내부 오류"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    options:
      tags: ["Voice Processing"]
      summary: "CORS preflight for transcription"
      description: "CORS preflight 요청을 처리합니다."
      operationId: "transcribeOptions"
      responses:
        '204':
          description: "CORS preflight 성공"
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: "*"
            Access-Control-Allow-Methods:
              schema:
                type: string
                example: "POST, OPTIONS"
            Access-Control-Allow-Headers:
              schema:
                type: string
                example: "Content-Type, Authorization"

  # Voice Entries
...